Add survey form with ajax content population
Survey creation will be hold in a form whose input fields will be contructed throughout the process by the user.
User can click the Add Question button to add a question field and Add Answer to add an answer field to the last question. When submit, the whole form will be sent to the server and save all the questions and corresponding answers to the database.